xhs-ai-auto/requirements.txt
requests
python-dotenv
google-genai
Pillow
